Cloud and ExtendedCloud extend from new EventDispatcher base class
They had the same implementation of the event dispatching logic.
Only seemed obvious this should be in a common base class.
